Make use of abstract UserTokensProvider class
That bout registered and unregistered user token provider classes can
use. Instead of one extends the other we can have one abstract "recipe"
of the classes.
